Move balanceCacheKey to its own file to satisify go fuzz build
